+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [] - Unreleased
 
+* Improve error logging when we fail to find the public ip address of a vm.
+
 ## [1.3.1] - 2018-05-10
 
 * Pin docker to docker-17.09.1ce-1.111.amzn1 on Amazon since the old package that we pinned to has disappeared.